+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Fixed
 
+- token replacement of {TARGET|GEM|NAME}
+
 ### Security
 
 ## [1.1.58] - 2025-11-13